+++ src/com/nixinova/main/Mineo.java
@@ -12,7 +12,7 @@ import com.nixinova.input.InputHandler;
 import com.nixinova.options.Options;
 
 public class Mineo {
-	public static final String VERSION = "0.0.16_1";
+	public static final String VERSION = "0.0.17";
 	public static final String TITLE = "Mineo " + VERSION;
 
 	private static JFrame frame;